Move uncommon track and road piece code to construction tab.
