+++ src/openloco/windows/construction/Common.cpp
@@ -24,65 +24,6 @@ using namespace openloco::map::tilemgr;
 
 namespace openloco::ui::windows::construction
 {
-    static loco_global<previewTrack[1], 0x004F6D44> _unk_4F6D44;
-    static loco_global<previewTrack[1], 0x004F6D4F> _unk_4F6D4F;
-    static loco_global<previewTrack[1], 0x004F6D5A> _unk_4F6D5A;
-    static loco_global<previewTrack[4], 0x004F6D65> _unk_4F6D65;
-    static loco_global<previewTrack[4], 0x004F6D8E> _unk_4F6D8E;
-    static loco_global<previewTrack[2], 0x004F6DB7> _unk_4F6DB7;
-    static loco_global<previewTrack[2], 0x004F6DCC> _unk_4F6DCC;
-    static loco_global<previewTrack[1], 0x004F6DE1> _unk_4F6DE1;
-    static loco_global<previewTrack[1], 0x004F6DEC> _unk_4F6DEC;
-    static loco_global<previewTrack[1], 0x004F6DF7> _unk_4F6DF7;
-
-    static loco_global<previewTrack[1], 0x004F7488> _unk_4F7488;
-    static loco_global<previewTrack[4], 0x004F7493> _unk_4F7493;
-    static loco_global<previewTrack[1], 0x004F74BC> _unk_4F74BC;
-    static loco_global<previewTrack[1], 0x004F74C7> _unk_4F74C7;
-    static loco_global<previewTrack[4], 0x004F74D2> _unk_4F74D2;
-    static loco_global<previewTrack[4], 0x004F74FB> _unk_4F74FB;
-    static loco_global<previewTrack[5], 0x004F7524> _unk_4F7524;
-    static loco_global<previewTrack[5], 0x004F7557> _unk_4F7557;
-    static loco_global<previewTrack[5], 0x004F758A> _unk_4F758A;
-    static loco_global<previewTrack[5], 0x004F75BD> _unk_4F75BD;
-    static loco_global<previewTrack[5], 0x004F75F0> _unk_4F75F0;
-    static loco_global<previewTrack[5], 0x004F7623> _unk_4F7623;
-    static loco_global<previewTrack[4], 0x004F7656> _unk_4F7656;
-    static loco_global<previewTrack[4], 0x004F767F> _unk_4F767F;
-    static loco_global<previewTrack[2], 0x004F76A8> _unk_4F76A8;
-    static loco_global<previewTrack[2], 0x004F76BD> _unk_4F76BD;
-    static loco_global<previewTrack[1], 0x004F76D2> _unk_4F76D2;
-    static loco_global<previewTrack[1], 0x004F76DD> _unk_4F76DD;
-    static loco_global<previewTrack[4], 0x004F76E8> _unk_4F76E8;
-    static loco_global<previewTrack[4], 0x004F7711> _unk_4F7711;
-    static loco_global<previewTrack[4], 0x004F773A> _unk_4F773A;
-    static loco_global<previewTrack[4], 0x004F7763> _unk_4F7763;
-    static loco_global<previewTrack[4], 0x004F778C> _unk_4F778C;
-    static loco_global<previewTrack[4], 0x004F77B5> _unk_4F77B5;
-    static loco_global<previewTrack[4], 0x004F77DE> _unk_4F77DE;
-    static loco_global<previewTrack[4], 0x004F7807> _unk_4F7807;
-    static loco_global<previewTrack[1], 0x004F7830> _unk_4F7830;
-    static loco_global<previewTrack[1], 0x004F783B> _unk_4F783B;
-    static loco_global<previewTrack[1], 0x004F7846> _unk_4F7846;
-    static loco_global<previewTrack[1], 0x004F7851> _unk_4F7851;
-    static loco_global<previewTrack[1], 0x004F785C> _unk_4F785C;
-    static loco_global<previewTrack[1], 0x004F7867> _unk_4F7867;
-    static loco_global<previewTrack[1], 0x004F7872> _unk_4F7872;
-    static loco_global<previewTrack[1], 0x004F787D> _unk_4F787D;
-    static loco_global<previewTrack[1], 0x004F7888> _unk_4F7888;
-    static loco_global<previewTrack[1], 0x004F7893> _unk_4F7893;
-    static loco_global<previewTrack[1], 0x004F789E> _unk_4F789E;
-    static loco_global<previewTrack[1], 0x004F78A9> _unk_4F78A9;
-    static loco_global<previewTrack[1], 0x004F78B4> _unk_4F78B4;
-    static loco_global<previewTrack[1], 0x004F78BF> _unk_4F78BF;
-    static loco_global<previewTrack[1], 0x004F78CA> _unk_4F78CA;
-    static loco_global<previewTrack[1], 0x004F78D5> _unk_4F78D5;
-    static loco_global<previewTrack[1], 0x004F78E0> _unk_4F78E0;
-    static loco_global<previewTrack[1], 0x004F78EB> _unk_4F78EB;
-
-    static loco_global<uint8_t, 0x00522090> _byte_522090;
-    static loco_global<uint8_t, 0x00522091> _byte_522091;
-    static loco_global<uint8_t, 0x00522092> _byte_522092;
     static loco_global<ui::window_number, 0x00523390> _toolWindowNumber;
 
     static window* nonTrackWindow()
@@ -112,7 +53,7 @@ namespace openloco::ui::windows::construction
             common::setDisabledWidgets(window);
         }
 
-        common::activateSelectedConstructionWidgets();
+        construction::activateSelectedConstructionWidgets();
         window = WindowManager::find(WindowType::construction);
 
         if (window != nullptr)
@@ -410,82 +351,6 @@ namespace openloco::ui::windows::construction
 
     namespace common
     {
-        const uint8_t trackPieceWidgets[] = {
-            construction::widx::straight,
-            construction::widx::left_hand_curve_very_small,
-            construction::widx::right_hand_curve_very_small,
-            construction::widx::left_hand_curve_small,
-            construction::widx::right_hand_curve_small,
-            construction::widx::left_hand_curve,
-            construction::widx::right_hand_curve,
-            construction::widx::left_hand_curve_large,
-            construction::widx::right_hand_curve_large,
-            construction::widx::s_bend_left,
-            construction::widx::s_bend_right,
-        };
-
-        // 0x004F6D1C
-        const previewTrack* roadPieces[] = {
-            _unk_4F6D44,
-            _unk_4F6D4F,
-            _unk_4F6D5A,
-            _unk_4F6D65,
-            _unk_4F6D8E,
-            _unk_4F6DB7,
-            _unk_4F6DCC,
-            _unk_4F6DE1,
-            _unk_4F6DEC,
-            _unk_4F6DF7,
-        };
-
-        // 0x004F73D8
-        const previewTrack* trackPieces[] = {
-            _unk_4F7488,
-            _unk_4F7493,
-            _unk_4F74BC,
-            _unk_4F74C7,
-            _unk_4F74D2,
-            _unk_4F74FB,
-            _unk_4F7524,
-            _unk_4F7557,
-            _unk_4F758A,
-            _unk_4F75BD,
-            _unk_4F75F0,
-            _unk_4F7623,
-            _unk_4F7656,
-            _unk_4F767F,
-            _unk_4F76A8,
-            _unk_4F76BD,
-            _unk_4F76D2,
-            _unk_4F76DD,
-            _unk_4F76E8,
-            _unk_4F7711,
-            _unk_4F773A,
-            _unk_4F7763,
-            _unk_4F778C,
-            _unk_4F77B5,
-            _unk_4F77DE,
-            _unk_4F7807,
-            _unk_4F7830,
-            _unk_4F783B,
-            _unk_4F7846,
-            _unk_4F7851,
-            _unk_4F785C,
-            _unk_4F7867,
-            _unk_4F7872,
-            _unk_4F787D,
-            _unk_4F7888,
-            _unk_4F7893,
-            _unk_4F789E,
-            _unk_4F78A9,
-            _unk_4F78B4,
-            _unk_4F78BF,
-            _unk_4F78CA,
-            _unk_4F78D5,
-            _unk_4F78E0,
-            _unk_4F78EB,
-        };
-
         struct TabInformation
         {
             widget_t* widgets;
@@ -530,7 +395,7 @@ namespace openloco::ui::windows::construction
 
             if (widgetIndex == widx::tab_construction)
             {
-                common::activateSelectedConstructionWidgets();
+                construction::activateSelectedConstructionWidgets();
             }
 
             common::sub_49FEC7();
@@ -861,852 +726,6 @@ namespace openloco::ui::windows::construction
             station::init_events();
             signal::init_events();
             overhead::init_events();
-        }
-
-        // 0x004A0832
-        std::optional<trackPieceId> getRoadPieceId(uint8_t trackPiece, uint8_t gradient, uint8_t rotation)
-        {
-            if (trackPiece == 0xFF)
-                return std::nullopt;
-
-            uint8_t id = 0;
-
-            switch (trackPiece)
-            {
-                case common::trackPiece::straight: // 0x004A0856
-                {
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    switch (gradient)
-                    {
-                        default:
-                            return std::nullopt;
-                        case common::trackGradient::level:
-                            id = 0;
-                            break;
-                        case common::trackGradient::slope_up:
-                            id = 5;
-                            break;
-                        case common::trackGradient::slope_down:
-                            id = 6;
-                            break;
-                        case common::trackGradient::steep_slope_up:
-                            id = 7;
-                            break;
-                        case common::trackGradient::steep_slope_down:
-                            id = 8;
-                            break;
-                    }
-                    break;
-                }
-                case common::trackPiece::left_hand_curve_very_small: // 0x004A08A5
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 1;
-                    break;
-                }
-                case common::trackPiece::right_hand_curve_very_small: // 0x004A08CD
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 2;
-                    break;
-                }
-                case common::trackPiece::left_hand_curve_small: // 0x004A08ED
-                {
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 3;
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    break;
-                }
-                case common::trackPiece::right_hand_curve_small: // 0x004A08FB
-                {
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 4;
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    break;
-                }
-                case common::trackPiece::left_hand_curve: // 0x004A095F
-                case common::trackPiece::right_hand_curve:
-                case common::trackPiece::left_hand_curve_large:
-                case common::trackPiece::right_hand_curve_large:
-                case common::trackPiece::s_bend_left:
-                case common::trackPiece::s_bend_right:
-                case common::trackPiece::s_bend_to_dual_track:
-                case common::trackPiece::s_bend_to_single_track:
-                {
-                    return std::nullopt;
-                }
-                case common::trackPiece::turnaround: // 0x004A0909
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 12)
-                        return std::nullopt;
-                    id = 9;
-                    break;
-                }
-            }
-
-            if (rotation < 12)
-                rotation &= 3;
-
-            return trackPieceId{ id, rotation };
-        }
-
-        // 0x004A04F8
-        std::optional<trackPieceId> getTrackPieceId(uint8_t trackPiece, uint8_t gradient, uint8_t rotation)
-        {
-            if (trackPiece == 0xFF)
-                return std::nullopt;
-
-            uint8_t id = 0;
-
-            switch (trackPiece)
-            {
-                case common::trackPiece::straight: // loc_4A051C
-                {
-                    if (rotation >= 12)
-                    {
-                        id = 1;
-                        if (gradient != common::trackGradient::level)
-                            return std::nullopt;
-                    }
-                    else
-                    {
-                        if (rotation >= 8)
-                        {
-                            switch (gradient)
-                            {
-                                default:
-                                    return std::nullopt;
-                                case common::trackGradient::level:
-                                    id = 27;
-                                    break;
-                                case common::trackGradient::steep_slope_up:
-                                    id = 35;
-                                    break;
-                                case common::trackGradient::steep_slope_down:
-                                    id = 37;
-                                    break;
-                            }
-                        }
-                        else if (rotation >= 4)
-                        {
-                            switch (gradient)
-                            {
-                                default:
-                                    return std::nullopt;
-                                case common::trackGradient::level:
-                                    id = 26;
-                                    break;
-                                case common::trackGradient::steep_slope_up:
-                                    id = 34;
-                                    break;
-                                case common::trackGradient::steep_slope_down:
-                                    id = 36;
-                                    break;
-                            }
-                        }
-                        else
-                        {
-                            switch (gradient)
-                            {
-                                default:
-                                    return std::nullopt;
-                                case common::trackGradient::level:
-                                    id = 0;
-                                    break;
-                                case common::trackGradient::slope_up:
-                                    id = 14;
-                                    break;
-                                case common::trackGradient::slope_down:
-                                    id = 15;
-                                    break;
-                                case common::trackGradient::steep_slope_up:
-                                    id = 16;
-                                    break;
-                                case common::trackGradient::steep_slope_down:
-                                    id = 17;
-                                    break;
-                            }
-                        }
-                    }
-                    break;
-                }
-
-                case common::trackPiece::left_hand_curve_very_small: // loc_4A05C3
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 12)
-                        return std::nullopt;
-                    if (rotation >= 8)
-                    {
-                        id = 29;
-                        break;
-                    }
-                    if (rotation >= 4)
-                    {
-                        id = 28;
-                        break;
-                    }
-                    id = 2;
-                    break;
-                }
-
-                case common::trackPiece::right_hand_curve_very_small: // loc_4A05F4
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 12)
-                        return std::nullopt;
-                    if (rotation >= 8)
-                    {
-                        id = 31;
-                        break;
-                    }
-                    if (rotation >= 4)
-                    {
-                        id = 30;
-                        break;
-                    }
-                    id = 3;
-                    break;
-                }
-
-                case common::trackPiece::left_hand_curve_small: // loc_4A0625
-                {
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    switch (gradient)
-                    {
-                        default:
-                            return std::nullopt;
-                        case common::trackGradient::level:
-                            id = 4;
-                            break;
-                        case common::trackGradient::slope_up:
-                            id = 18;
-                            break;
-                        case common::trackGradient::slope_down:
-                            id = 20;
-                            break;
-                        case common::trackGradient::steep_slope_up:
-                            id = 22;
-                            break;
-                        case common::trackGradient::steep_slope_down:
-                            id = 24;
-                            break;
-                    }
-                    break;
-                }
-
-                case common::trackPiece::right_hand_curve_small: // loc_4A066A
-                {
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    switch (gradient)
-                    {
-                        default:
-                            return std::nullopt;
-                        case common::trackGradient::level:
-                            id = 5;
-                            break;
-                        case common::trackGradient::slope_up:
-                            id = 19;
-                            break;
-                        case common::trackGradient::slope_down:
-                            id = 21;
-                            break;
-                        case common::trackGradient::steep_slope_up:
-                            id = 23;
-                            break;
-                        case common::trackGradient::steep_slope_down:
-                            id = 25;
-                            break;
-                    }
-                    break;
-                }
-
-                case common::trackPiece::left_hand_curve: // loc_4A06AF
-                {
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 6;
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    break;
-                }
-
-                case common::trackPiece::right_hand_curve: // loc_4A06C8
-                {
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 7;
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    break;
-                }
-
-                case common::trackPiece::left_hand_curve_large: // loc_4A06E1
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    id = 10;
-                    if (rotation >= 12)
-                        break;
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 8;
-                    break;
-                }
-
-                case common::trackPiece::right_hand_curve_large: // loc_4A0705
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    id = 11;
-                    if (rotation >= 12)
-                        break;
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 9;
-                    break;
-                }
-
-                case common::trackPiece::s_bend_left: // loc_4A0729
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 12)
-                        return std::nullopt;
-                    id = 33;
-                    if (rotation >= 8)
-                        break;
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 12;
-                    break;
-                }
-
-                case common::trackPiece::s_bend_right: // loc_4A0756
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 8)
-                        return std::nullopt;
-                    id = 32;
-                    if (rotation >= 4)
-                        break;
-                    id = 13;
-                    break;
-                }
-
-                case common::trackPiece::s_bend_to_dual_track: // loc_4A077C
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 8)
-                        return std::nullopt;
-                    id = 40;
-                    if (rotation >= 4)
-                        break;
-                    id = 38;
-                    break;
-                }
-
-                case common::trackPiece::s_bend_to_single_track: // loc_4A07A2
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 12)
-                        return std::nullopt;
-                    id = 41;
-                    if (rotation >= 8)
-                        break;
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    id = 39;
-                    break;
-                }
-
-                case common::trackPiece::turnaround: // loc_4A07C0
-                {
-                    if (gradient != common::trackGradient::level)
-                        return std::nullopt;
-                    if (rotation >= 12)
-                        return std::nullopt;
-                    id = 43;
-                    if (rotation >= 8)
-                        break;
-                    id = 42;
-                    if (rotation >= 4)
-                        return std::nullopt;
-                    break;
-                }
-            }
-
-            if (rotation < 12)
-                rotation &= 3;
-
-            return trackPieceId{ id, rotation };
-        }
-
-        static void activateSelectedRoadWidgets(window* window)
-        {
-            tilemgr::map_invalidate_map_selection_tiles();
-            _mapSelectionFlags = _mapSelectionFlags | (1 << 3) | (1 << 1);
-
-            auto road = getRoadPieceId(_lastSelectedTrackPiece, _lastSelectedTrackGradient, _constructionRotation);
-
-            uint8_t rotation;
-            uint8_t roadId;
-
-            auto x = _x;
-            auto y = _y;
-
-            if (!road)
-            {
-                rotation = _constructionRotation;
-                roadId = 0;
-            }
-            else
-            {
-                rotation = road->rotation;
-                roadId = road->id;
-            }
-
-            auto roadPiece = roadPieces[roadId];
-            auto i = 0;
-            auto posId = 0;
-            rotation &= 3;
-
-            while (roadPiece[i].index != 0xFF)
-            {
-                if (roadPiece[i].flags & previewTrackFlags::diagonal)
-                {
-                    i++;
-                    continue;
-                }
-
-                map_pos pos = { roadPiece[i].x, roadPiece[i].y };
-
-                pos = rotate2DCoordinate(pos, rotation);
-
-                pos.x += x;
-                pos.y += y;
-                _mapSelectedTiles[posId] = pos;
-                posId++;
-                i++;
-            }
-
-            _mapSelectedTiles[posId].x = -1;
-            map_invalidate_map_selection_tiles();
-            window->holdable_widgets = (1 << construction::widx::construct) | (1 << construction::widx::remove);
-
-            auto trackType = _trackType & ~(1 << 7);
-            auto roadObj = objectmgr::get<road_object>(trackType);
-
-            window->widgets[construction::widx::s_bend_left].type = widget_type::none;
-            window->widgets[construction::widx::s_bend_right].type = widget_type::none;
-            window->widgets[construction::widx::left_hand_curve_large].type = widget_type::none;
-            window->widgets[construction::widx::right_hand_curve_large].type = widget_type::none;
-            window->widgets[construction::widx::left_hand_curve].type = widget_type::none;
-            window->widgets[construction::widx::right_hand_curve].type = widget_type::none;
-            window->widgets[construction::widx::left_hand_curve_small].type = widget_type::none;
-            window->widgets[construction::widx::right_hand_curve_small].type = widget_type::none;
-            window->widgets[construction::widx::left_hand_curve_very_small].type = widget_type::none;
-            window->widgets[construction::widx::right_hand_curve_very_small].type = widget_type::none;
-
-            window->widgets[construction::widx::left_hand_curve_small].left = 3;
-            window->widgets[construction::widx::left_hand_curve_small].right = 24;
-            window->widgets[construction::widx::right_hand_curve_small].left = 113;
-            window->widgets[construction::widx::right_hand_curve_small].right = 134;
-            window->widgets[construction::widx::left_hand_curve].left = 25;
-            window->widgets[construction::widx::left_hand_curve].right = 46;
-            window->widgets[construction::widx::right_hand_curve].left = 91;
-            window->widgets[construction::widx::right_hand_curve].right = 112;
-
-            if (roadObj->road_pieces & road_piece_flags::track)
-            {
-                window->widgets[construction::widx::left_hand_curve_small].left = 25;
-                window->widgets[construction::widx::left_hand_curve_small].right = 46;
-                window->widgets[construction::widx::right_hand_curve_small].left = 91;
-                window->widgets[construction::widx::right_hand_curve_small].right = 112;
-                window->widgets[construction::widx::left_hand_curve].left = 47;
-                window->widgets[construction::widx::left_hand_curve].right = 68;
-                window->widgets[construction::widx::right_hand_curve].left = 69;
-                window->widgets[construction::widx::right_hand_curve].right = 90;
-
-                window->widgets[construction::widx::left_hand_curve_very_small].type = widget_type::wt_9;
-                window->widgets[construction::widx::right_hand_curve_very_small].type = widget_type::wt_9;
-            }
-
-            if (roadObj->road_pieces & road_piece_flags::one_way)
-            {
-                window->widgets[construction::widx::left_hand_curve_small].type = widget_type::wt_9;
-                window->widgets[construction::widx::right_hand_curve_small].type = widget_type::wt_9;
-            }
-
-            window->widgets[construction::widx::s_bend_dual_track_left].type = widget_type::none;
-            window->widgets[construction::widx::s_bend_dual_track_right].type = widget_type::none;
-
-            if (roadObj->road_pieces & road_piece_flags::one_sided)
-            {
-                window->widgets[construction::widx::s_bend_dual_track_left].type = widget_type::wt_9;
-                window->widgets[construction::widx::s_bend_dual_track_left].image = image_ids::construction_right_turnaround;
-                window->widgets[construction::widx::s_bend_dual_track_left].tooltip = string_ids::tooltip_turnaround;
-
-                if (_byte_525FAE == 0)
-                    window->widgets[construction::widx::s_bend_dual_track_left].image = image_ids::construction_left_turnaround;
-            }
-
-            window->widgets[construction::widx::steep_slope_down].type = widget_type::none;
-            window->widgets[construction::widx::slope_down].type = widget_type::none;
-            window->widgets[construction::widx::slope_up].type = widget_type::none;
-            window->widgets[construction::widx::steep_slope_up].type = widget_type::none;
-
-            if (roadObj->road_pieces & road_piece_flags::slope)
-            {
-                window->widgets[construction::widx::slope_down].type = widget_type::wt_9;
-                window->widgets[construction::widx::slope_up].type = widget_type::wt_9;
-            }
-
-            if (roadObj->road_pieces & road_piece_flags::steep_slope)
-            {
-                window->widgets[construction::widx::steep_slope_down].type = widget_type::wt_9;
-                window->widgets[construction::widx::steep_slope_up].type = widget_type::wt_9;
-            }
-
-            window->widgets[construction::widx::bridge].type = widget_type::wt_18;
-            window->widgets[construction::widx::bridge_dropdown].type = widget_type::wt_11;
-
-            if (_lastSelectedBridge == 0xFF || (_constructionHover != 1 && !(_byte_1136076 & 1)))
-            {
-                window->widgets[construction::widx::bridge].type = widget_type::none;
-                window->widgets[construction::widx::bridge_dropdown].type = widget_type::none;
-            }
-
-            auto activatedWidgets = window->activated_widgets;
-            activatedWidgets &= ~(construction::allTrack);
-
-            window->widgets[construction::widx::construct].type = widget_type::none;
-            window->widgets[construction::widx::remove].type = widget_type::wt_9;
-            window->widgets[construction::widx::rotate_90].type = widget_type::none;
-
-            if (_constructionHover == 1)
-            {
-                window->widgets[construction::widx::construct].type = widget_type::wt_5;
-                window->widgets[construction::widx::construct].tooltip = string_ids::tooltip_start_construction;
-                window->widgets[construction::widx::remove].type = widget_type::none;
-                window->widgets[construction::widx::rotate_90].type = widget_type::wt_9;
-                window->widgets[construction::widx::rotate_90].image = image_ids::rotate_object;
-                window->widgets[construction::widx::rotate_90].tooltip = string_ids::rotate_90;
-            }
-            else if (_constructionHover == 0)
-            {
-                window->widgets[construction::widx::construct].type = widget_type::wt_3;
-                window->widgets[construction::widx::construct].tooltip = string_ids::tooltip_construct;
-                window->widgets[construction::widx::rotate_90].type = widget_type::wt_9;
-                window->widgets[construction::widx::rotate_90].image = image_ids::construction_new_position;
-                window->widgets[construction::widx::rotate_90].tooltip = string_ids::new_construction_position;
-            }
-            if (_constructionHover == 0 || _constructionHover == 1)
-            {
-                if (_lastSelectedTrackPiece != 0xFF)
-                {
-                    auto trackPieceWidget = trackPieceWidgets[_lastSelectedTrackPiece];
-                    activatedWidgets |= 1ULL << trackPieceWidget;
-                }
-
-                uint8_t trackGradient = construction::widx::level;
-
-                switch (_lastSelectedTrackGradient)
-                {
-                    case common::trackGradient::level:
-                        trackGradient = construction::widx::level;
-                        break;
-
-                    case common::trackGradient::slope_up:
-                        trackGradient = construction::widx::slope_up;
-                        break;
-
-                    case common::trackGradient::slope_down:
-                        trackGradient = construction::widx::slope_down;
-                        break;
-
-                    case common::trackGradient::steep_slope_up:
-                        trackGradient = construction::widx::steep_slope_up;
-                        break;
-
-                    case common::trackGradient::steep_slope_down:
-                        trackGradient = construction::widx::steep_slope_down;
-                        break;
-                }
-
-                activatedWidgets |= 1ULL << trackGradient;
-            }
-            window->activated_widgets = activatedWidgets;
-            window->invalidate();
-        }
-
-        static void activateSelectedTrackWidgets(window* window)
-        {
-            tilemgr::map_invalidate_map_selection_tiles();
-            _mapSelectionFlags = _mapSelectionFlags | (1 << 3) | (1 << 1);
-
-            auto track = getTrackPieceId(_lastSelectedTrackPiece, _lastSelectedTrackGradient, _constructionRotation);
-
-            uint8_t rotation;
-            uint8_t trackId;
-            auto x = _x;
-            auto y = _y;
-
-            if (!track)
-            {
-                rotation = _constructionRotation;
-                trackId = 0;
-            }
-            else
-            {
-                rotation = track->rotation;
-                trackId = track->id;
-            }
-
-            auto trackPiece = trackPieces[trackId];
-            auto i = 0;
-            auto posId = 0;
-            rotation &= 3;
-
-            while (trackPiece[i].index != 0xFF)
-            {
-                if (trackPiece[i].flags & previewTrackFlags::diagonal)
-                {
-                    i++;
-                    continue;
-                }
-                map_pos pos = { trackPiece[i].x, trackPiece[i].y };
-
-                pos = rotate2DCoordinate(pos, rotation);
-
-                pos.x += x;
-                pos.y += y;
-                _mapSelectedTiles[posId] = pos;
-                posId++;
-                i++;
-            }
-
-            _mapSelectedTiles[posId].x = -1;
-            map_invalidate_map_selection_tiles();
-            window->holdable_widgets = (1 << construction::widx::construct) | (1 << construction::widx::remove);
-
-            auto trackObj = objectmgr::get<track_object>(_trackType);
-
-            window->widgets[construction::widx::s_bend_left].type = widget_type::wt_9;
-            window->widgets[construction::widx::s_bend_right].type = widget_type::wt_9;
-            window->widgets[construction::widx::left_hand_curve_large].type = widget_type::none;
-            window->widgets[construction::widx::right_hand_curve_large].type = widget_type::none;
-            window->widgets[construction::widx::left_hand_curve].type = widget_type::none;
-            window->widgets[construction::widx::right_hand_curve].type = widget_type::none;
-            window->widgets[construction::widx::left_hand_curve_small].type = widget_type::none;
-            window->widgets[construction::widx::right_hand_curve_small].type = widget_type::none;
-            window->widgets[construction::widx::left_hand_curve_very_small].type = widget_type::none;
-            window->widgets[construction::widx::right_hand_curve_very_small].type = widget_type::none;
-
-            window->widgets[construction::widx::left_hand_curve_small].left = 3;
-            window->widgets[construction::widx::left_hand_curve_small].right = 24;
-            window->widgets[construction::widx::right_hand_curve_small].left = 113;
-            window->widgets[construction::widx::right_hand_curve_small].right = 134;
-            window->widgets[construction::widx::left_hand_curve].left = 25;
-            window->widgets[construction::widx::left_hand_curve].right = 46;
-            window->widgets[construction::widx::right_hand_curve].left = 91;
-            window->widgets[construction::widx::right_hand_curve].right = 112;
-
-            if (trackObj->track_pieces & track_piece_flags::very_small_curve)
-            {
-                window->widgets[construction::widx::left_hand_curve_small].left = 25;
-                window->widgets[construction::widx::left_hand_curve_small].right = 46;
-                window->widgets[construction::widx::right_hand_curve_small].left = 91;
-                window->widgets[construction::widx::right_hand_curve_small].right = 112;
-                window->widgets[construction::widx::left_hand_curve].left = 47;
-                window->widgets[construction::widx::left_hand_curve].right = 68;
-                window->widgets[construction::widx::right_hand_curve].left = 69;
-                window->widgets[construction::widx::right_hand_curve].right = 90;
-
-                window->widgets[construction::widx::left_hand_curve_very_small].type = widget_type::wt_9;
-                window->widgets[construction::widx::right_hand_curve_very_small].type = widget_type::wt_9;
-            }
-
-            if (trackObj->track_pieces & track_piece_flags::large_curve)
-            {
-                window->widgets[construction::widx::left_hand_curve_large].type = widget_type::wt_9;
-                window->widgets[construction::widx::right_hand_curve_large].type = widget_type::wt_9;
-            }
-
-            if (trackObj->track_pieces & track_piece_flags::normal_curve)
-            {
-                window->widgets[construction::widx::left_hand_curve].type = widget_type::wt_9;
-                window->widgets[construction::widx::right_hand_curve].type = widget_type::wt_9;
-            }
-
-            if (trackObj->track_pieces & track_piece_flags::small_curve)
-            {
-                window->widgets[construction::widx::left_hand_curve_small].type = widget_type::wt_9;
-                window->widgets[construction::widx::right_hand_curve_small].type = widget_type::wt_9;
-            }
-
-            window->widgets[construction::widx::s_bend_dual_track_left].type = widget_type::none;
-            window->widgets[construction::widx::s_bend_dual_track_right].type = widget_type::none;
-
-            if (trackObj->track_pieces & track_piece_flags::one_sided)
-            {
-                window->widgets[construction::widx::s_bend_dual_track_left].type = widget_type::wt_9;
-                window->widgets[construction::widx::s_bend_dual_track_right].type = widget_type::wt_9;
-                window->widgets[construction::widx::s_bend_dual_track_left].image = image_ids::construction_s_bend_dual_track_left;
-                window->widgets[construction::widx::s_bend_dual_track_right].image = image_ids::construction_s_bend_dual_track_right;
-                window->widgets[construction::widx::s_bend_dual_track_left].tooltip = string_ids::tooltip_s_bend_left_dual_track;
-                window->widgets[construction::widx::s_bend_dual_track_right].tooltip = string_ids::tooltip_s_bend_right_dual_track;
-
-                _byte_522090 = 16;
-                _byte_522091 = 20;
-
-                if (_constructionRotation >= 4 && _constructionRotation < 12)
-                {
-                    window->widgets[construction::widx::s_bend_dual_track_left].image = image_ids::construction_right_turnaround;
-                    window->widgets[construction::widx::s_bend_dual_track_right].image = image_ids::construction_s_bend_to_single_track_left;
-                    window->widgets[construction::widx::s_bend_dual_track_left].tooltip = string_ids::tooltip_turnaround;
-                    window->widgets[construction::widx::s_bend_dual_track_right].tooltip = string_ids::tooltip_s_bend_to_single_track;
-                    _byte_522090 = 20;
-                    _byte_522092 = 16;
-                    if (_constructionRotation >= 8)
-                    {
-                        window->widgets[construction::widx::s_bend_dual_track_left].image = image_ids::construction_s_bend_to_single_track_right;
-                        window->widgets[construction::widx::s_bend_dual_track_right].image = image_ids::construction_left_turnaround;
-                        window->widgets[construction::widx::s_bend_dual_track_left].tooltip = string_ids::tooltip_s_bend_to_single_track;
-                        window->widgets[construction::widx::s_bend_dual_track_right].tooltip = string_ids::tooltip_turnaround;
-                        _byte_522091 = 16;
-                        _byte_522092 = 20;
-                    }
-                }
-            }
-            window->widgets[construction::widx::steep_slope_down].type = widget_type::none;
-            window->widgets[construction::widx::slope_down].type = widget_type::none;
-            window->widgets[construction::widx::slope_up].type = widget_type::none;
-            window->widgets[construction::widx::steep_slope_up].type = widget_type::none;
-
-            if (trackObj->track_pieces & track_piece_flags::slope)
-            {
-                window->widgets[construction::widx::slope_down].type = widget_type::wt_9;
-                window->widgets[construction::widx::slope_up].type = widget_type::wt_9;
-            }
-
-            if (trackObj->track_pieces & track_piece_flags::steep_slope)
-            {
-                window->widgets[construction::widx::steep_slope_down].type = widget_type::wt_9;
-                window->widgets[construction::widx::steep_slope_up].type = widget_type::wt_9;
-            }
-
-            window->widgets[construction::widx::bridge].type = widget_type::wt_18;
-            window->widgets[construction::widx::bridge_dropdown].type = widget_type::wt_11;
-
-            if (_lastSelectedBridge == 0xFF || (_constructionHover != 1 && !(_byte_1136076 & 1)))
-            {
-                window->widgets[construction::widx::bridge].type = widget_type::none;
-                window->widgets[construction::widx::bridge_dropdown].type = widget_type::none;
-            }
-
-            auto activatedWidgets = window->activated_widgets;
-            activatedWidgets &= ~(construction::allTrack);
-
-            window->widgets[construction::widx::construct].type = widget_type::none;
-            window->widgets[construction::widx::remove].type = widget_type::wt_9;
-            window->widgets[construction::widx::rotate_90].type = widget_type::none;
-
-            if (_constructionHover == 1)
-            {
-                window->widgets[construction::widx::construct].type = widget_type::wt_5;
-                window->widgets[construction::widx::construct].tooltip = string_ids::tooltip_start_construction;
-                window->widgets[construction::widx::remove].type = widget_type::none;
-                window->widgets[construction::widx::rotate_90].type = widget_type::wt_9;
-                window->widgets[construction::widx::rotate_90].image = image_ids::rotate_object;
-                window->widgets[construction::widx::rotate_90].tooltip = string_ids::rotate_90;
-            }
-            else if (_constructionHover == 0)
-            {
-                window->widgets[construction::widx::construct].type = widget_type::wt_3;
-                window->widgets[construction::widx::construct].tooltip = string_ids::tooltip_construct;
-                window->widgets[construction::widx::rotate_90].type = widget_type::wt_9;
-                window->widgets[construction::widx::rotate_90].image = image_ids::construction_new_position;
-                window->widgets[construction::widx::rotate_90].tooltip = string_ids::new_construction_position;
-            }
-            if (_constructionHover == 0 || _constructionHover == 1)
-            {
-                if (_lastSelectedTrackPiece != 0xFF)
-                {
-                    auto trackPieceWidget = trackPieceWidgets[_lastSelectedTrackPiece];
-                    activatedWidgets |= 1ULL << trackPieceWidget;
-                }
-
-                uint8_t trackGradient = construction::widx::level;
-
-                switch (_lastSelectedTrackGradient)
-                {
-                    case common::trackGradient::level:
-                        trackGradient = construction::widx::level;
-                        break;
-
-                    case common::trackGradient::slope_up:
-                        trackGradient = construction::widx::slope_up;
-                        break;
-
-                    case common::trackGradient::slope_down:
-                        trackGradient = construction::widx::slope_down;
-                        break;
-
-                    case common::trackGradient::steep_slope_up:
-                        trackGradient = construction::widx::steep_slope_up;
-                        break;
-
-                    case common::trackGradient::steep_slope_down:
-                        trackGradient = construction::widx::steep_slope_down;
-                        break;
-                }
-
-                activatedWidgets |= 1ULL << trackGradient;
-            }
-            window->activated_widgets = activatedWidgets;
-            window->invalidate();
-        }
-
-        // 0x0049F1B5
-        void activateSelectedConstructionWidgets()
-        {
-            auto window = WindowManager::find(WindowType::construction);
-
-            if (window == nullptr)
-                return;
-
-            if (window->current_tab == widx::tab_construction - widx::tab_construction)
-            {
-                if (_trackType & (1 << 7))
-                {
-                    activateSelectedRoadWidgets(window);
-                }
-                else
-                {
-                    activateSelectedTrackWidgets(window);
-                }
-            }
         }
 
         // 0x004CD454
@@ -2159,7 +1178,7 @@ namespace openloco::ui::windows::construction
             {
                 setDisabledWidgets(window);
             }
-            common::activateSelectedConstructionWidgets();
+            construction::activateSelectedConstructionWidgets();
         }
 
         // 0x00488B4D
@@ -2224,7 +1243,7 @@ namespace openloco::ui::windows::construction
             0x0049F1B5,
             [](registers& regs) FORCE_ALIGN_ARG_POINTER -> uint8_t {
                 registers backup = regs;
-                common::activateSelectedConstructionWidgets();
+                construction::activateSelectedConstructionWidgets();
                 regs = backup;
                 return 0;
             });
